Rename receiver package name.
