+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/DispatcherManager.java
@@ -20,7 +20,7 @@ package org.apache.skywalking.oap.server.core.analysis;
 
 import java.util.*;
 import org.apache.skywalking.oap.server.core.analysis.endpoint.EndpointDispatcher;
-import org.apache.skywalking.oap.server.core.receiver.Scope;
+import org.apache.skywalking.oap.server.core.source.Scope;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.slf4j.*;
