+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/SourceReceiver.java
@@ -16,28 +16,13 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.receiver;
+package org.apache.skywalking.oap.server.core.source;
 
-import lombok.*;
-import org.apache.skywalking.apm.network.language.agent.*;
+import org.apache.skywalking.oap.server.library.module.Service;
 
 /**
  * @author peng-yongsheng
  */
-public class EndpointRelation extends Source {
-
-    @Override public Scope scope() {
-        return Scope.EndpointRelation;
-    }
-
-    @Getter @Setter private int endpointId;
-    @Getter @Setter private String endpoint;
-    @Getter @Setter private int childEndpointId;
-    @Getter @Setter private String childEndpoint;
-    @Getter @Setter private int rpcLatency;
-    @Getter @Setter private boolean status;
-    @Getter @Setter private int responseCode;
-    @Getter @Setter private SpanLayer type;
-    @Getter @Setter private SpanType detectPoint;
+public interface SourceReceiver extends Service {
+    void receive(Source source);
 }
-
